rewrote tests
